added server side function for api get calls
